# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Management System\n# Error codes\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e120001\u003c/code\u003e: Wrong pagination query\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e120002\u003c/code\u003e: Wrong query\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e120003\u003c/code\u003e: Wrong body\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e121101\u003c/code\u003e: Email template not found\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e121204\u003c/code\u003e: Invalid order\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e121205\u003c/code\u003e: Email template ID is missing\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e121206\u003c/code\u003e: Invalid email template ID\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e121207\u003c/code\u003e: Email template not found\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e121208\u003c/code\u003e: Stage not found\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e121301\u003c/code\u003e: Reject reason not found\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e121401\u003c/code\u003e: Question set not found\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e121501\u003c/code\u003e: Hiring team not found\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e121502\u003c/code\u003e: User not found\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e122101\u003c/code\u003e: Position not found\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e122201\u003c/code\u003e: Required field\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e122202\u003c/code\u003e: Stage not found\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e122203\u003c/code\u003e: Black list or duplicate\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e122204\u003c/code\u003e: Pool not found\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e122205\u003c/code\u003e: Candidate not found\u003c/span\u003e\n### \u003cspan style=\"font-size:18px\"\u003e- \u003ccode\u003e122301\u003c/code\u003e: Label not found\u003c/span\u003e",
        "title": "Management System API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/interfaces.TokenPair"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Logout",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Refresh token",
                        "name": "refreshToken",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Refresh token",
                        "name": "refreshToken",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/interfaces.TokenPair"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "Register",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.RegisterInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/roles": {
            "get": {
                "description": "Lấy danh sách roles với phân trang và filter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "List roles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by role name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Trang hiện tại (bắt đầu từ 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Số bản ghi mỗi trang",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ items: Role[], meta: { count,current_page,per_page,total,total_pages } }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Tạo role mới",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Create role",
                "parameters": [
                    {
                        "description": "Role payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Role"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Role"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/roles/{id}": {
            "get": {
                "description": "Lấy thông tin role theo ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Get role by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Role"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Cập nhật thông tin role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Update role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Role"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Role"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Xóa role theo ID",
                "tags": [
                    "roles"
                ],
                "summary": "Delete role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "interfaces.LoginInput": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "interfaces.RegisterInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "interfaces.TokenPair": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "model.Role": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "isSystem": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "permissionKeys": {
                    "description": "ví dụ: [\"employee.read\",\"employee.write\"]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}